+++ Source/Scene/Scene.js
@@ -105,6 +105,15 @@ define([
         this.skyBox = undefined;
 
         /**
+         * The sky atmosphere drawn around the globe.
+         *
+         * @type SkyAtmosphere
+         *
+         * @default undefined
+         */
+        this.skyAtmosphere = undefined;
+
+        /**
          * The current mode of the scene.
          *
          * @type SceneMode
@@ -402,17 +411,22 @@ define([
         var context = scene._context;
         var us = context.getUniformState();
         var skyBoxCommand = (typeof scene.skyBox !== 'undefined') ? scene.skyBox.update(context, scene._frameState) : undefined;
+        var skyAtmosphereCommand = (typeof scene.skyAtmosphere !== 'undefined') ? scene.skyAtmosphere.update(context, scene._frameState) : undefined;
 
         scene._clearColorCommand.execute(context, framebuffer);
 
-        if (typeof skyBoxCommand !== 'undefined') {
-            frustum.near = camera.frustum.near;
-            frustum.far = camera.frustum.far;
-            us.updateFrustum(frustum);
+        // Ideally, we would render the sky box and atmosphere last for
+        // early-z, but we would have to draw it in each frustum
+        frustum.near = camera.frustum.near;
+        frustum.far = camera.frustum.far;
+        us.updateFrustum(frustum);
 
-            // Ideally, we would render the sky box last for early-z, but
-            // we would have to draw it in each frustum
+        if (typeof skyBoxCommand !== 'undefined') {
             skyBoxCommand.execute(context, framebuffer);
+        }
+
+        if (typeof skyAtmosphereCommand !== 'undefined') {
+            skyAtmosphereCommand.execute(context, framebuffer);
         }
 
         var clearDepthStencil = scene._clearDepthStencilCommand;
@@ -647,6 +661,7 @@ define([
         this._pickFramebuffer = this._pickFramebuffer && this._pickFramebuffer.destroy();
         this._primitives = this._primitives && this._primitives.destroy();
         this.skyBox = this.skyBox && this.skyBox.destroy();
+        this.skyAtmosphere = this.skyAtmosphere && this.skyAtmosphere.destroy();
         this._context = this._context && this._context.destroy();
         return destroyObject(this);
     };